requirements.txt
torchvision
efficientnet_pytorch
tqdm
Pillow